refactored to support new addPortfolio interface adapter
